+++ frontend/src/Cabinet/types/dto/oAuth.dto.ts
@@ -1,8 +1,6 @@
 import { TOAuthProvider } from "@/Cabinet/assets/data/oAuth";
 
-// TODO : 주석 수정 필요. 연결 계정 정보 타입 정의
 export interface IUserOAuthLinkInfoDto {
   providerType: TOAuthProvider;
   email: string;
 }
-// TODO : login.dto -> oAuth.dto로 변경
