+++ frontend/src/Cabinet/utils/oAuthUtils.ts
@@ -13,4 +13,3 @@ export const getOAuthDisplayInfo = (
 ): IOAuthDisplay => {
   return OAUTH_CONFIG[provider].display;
 };
-// TODO : loginUtils 파일명 변경
